After spending some more time on that one weird part of the code where the LocalVertexCoordinates are multiplied by (0.01/2.55), I can say this:
This multiplication affects the resulting scale of the mesh. Because the mesh will stretch if needed the changes on scale are not noted inside the Render Chunks, but in between them, where the vertices do not share their indexes, suddenly a wild gap appears! After some tweaking I can say that 0.01/2.55 (not 2.5) is the correct number to be used. What's the difference between using 2.5 and 2.55? Neither of them makes the gap appear, like 2.6 does. Well, 2.5 works the other way around, making the scale smaller than what it should actually be, and the consecuence is that now you cannot choose between cubes with the cursor correctly. But 2.55 solves this, it works like a charm.